Move Slack Init
