+++ src/api/getUser/index.ts
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/node';
 
 import { SLACK_PROFILE_ID_GITHUB } from '@/config';
-import { bolt } from '@api/slack';
+import { bolt } from '@/init/slack';
 import { db } from '@utils/db';
 import { findUser } from '@utils/db/findUser';
 import { isSentrySlackUser } from '@utils/isSentrySlackUser';
