+++ src/brain/appHome/index.ts
@@ -1,6 +1,6 @@
+import { bolt } from '@/init/slack';
+import { updateAppHome } from '@/init/slack/updateAppHome';
 import { getUser } from '@api/getUser';
-import { bolt } from '@api/slack';
-import { updateAppHome } from '@api/slack/updateAppHome';
 import { db } from '@utils/db';
 import { normalizeGithubUser } from '@utils/normalizeGithubUser';
