+++ src/brain/ghaCancel/index.ts
@@ -6,7 +6,7 @@ import {
 } from '@slack/bolt';
 
 import { GH_ORGS } from '@/config';
-import { bolt } from '@api/slack';
+import { bolt } from '@/init/slack';
 
 async function handler({ event, say, client }) {
   // eslint-disable-next-line no-useless-escape
