+++ src/brain/gocdSlackFeeds/deployFeed.ts
@@ -11,10 +11,10 @@ import { WebAPICallResult } from '@slack/web-api';
 import { SlackMessageRow } from 'knex/types/tables';
 
 import { getUser } from '@/api/getUser';
-import { bolt } from '@/api/slack';
 import * as slackblocks from '@/blocks/slackBlocks';
 import { GETSENTRY_REPO_SLUG, GH_ORGS, SENTRY_REPO_SLUG } from '@/config';
 import { SlackMessage } from '@/config/slackMessage';
+import { bolt } from '@/init/slack';
 import { GoCDModification, GoCDPipeline, GoCDResponse } from '@/types';
 import { getLastGetSentryGoCDDeploy } from '@/utils/db/getLatestDeploy';
 import { getSlackMessage } from '@/utils/db/getSlackMessage';
