+++ src/brain/notifyOnGoCDStageEvent/index.test.ts
@@ -12,9 +12,9 @@ import {
   GOCD_SENTRYIO_FE_PIPELINE_NAME,
   REQUIRED_CHECK_NAME,
 } from '@/config';
+import { bolt } from '@/init/slack';
 import { Fastify } from '@/types';
 import { FINAL_STAGE_NAMES, INPROGRESS_MSG } from '@/utils/gocdHelpers';
-import { bolt } from '@api/slack';
 import { db } from '@utils/db';
 import * as metrics from '@utils/metrics';
