+++ src/brain/syncUserProfileChange/index.ts
@@ -1,5 +1,5 @@
 import { SLACK_PROFILE_ID_GITHUB } from '@/config';
-import { bolt } from '@api/slack';
+import { bolt } from '@/init/slack';
 import { db } from '@utils/db';
 import { isSentrySlackUser } from '@utils/isSentrySlackUser';
 import { normalizeGithubUser } from '@utils/normalizeGithubUser';
