+++ src/brain/typescript/index.test.ts
@@ -1,12 +1,12 @@
 import { createSlackAppMention } from '@test/utils/createSlackAppMention';
 
 import { buildServer } from '@/buildServer';
-import { bolt } from '@api/slack';
+import { bolt } from '@/init/slack';
 
 import getProgress from './getProgress';
 import { typescript } from '.';
 
-jest.mock('@api/slack');
+jest.mock('@/init/slack');
 
 jest.mock('./getProgress', () =>
   jest.fn(() => ({
