+++ src/brain/typescript/index.ts
@@ -1,5 +1,5 @@
 import { GETSENTRY_REPO_SLUG, SENTRY_REPO_SLUG } from '@/config';
-import { bolt } from '@api/slack';
+import { bolt } from '@/init/slack';
 import { wrapHandler } from '@utils/wrapHandler';
 
 import getProgress from './getProgress';
