+++ test/jest.setup.ts
@@ -1,4 +1,4 @@
 /* eslint-env jest */
-jest.mock('@api/slack');
+jest.mock('@/init/slack');
 jest.mock('@api/github/octokitWithRetries');
 jest.mock('@utils/loadBrain');
